Added tests and got examples working with script
