+++ pyRpc/__init__.py
@@ -42,12 +42,12 @@ POSSIBILITY OF SUCH DAMAGE.
 
 """
 
-from pyRpc.server import PyRpc, RpcResponse
-from pyRpc.client import RpcConnection, RpcRequest 
+#from pyRpc.server import PyRpc, RpcResponse
+#from pyRpc.client import RpcConnection, RpcRequest
 
-__version__ = '0.2'
+#__version__ = '0.2'
 
-__all__ = ['server', 'client', 'PyRpc', 'RpcResponse', 'RpcConnection', 'RpcRequest']
+#__all__ = ['server', 'client', 'PyRpc', 'RpcResponse', 'RpcConnection', 'RpcRequest']
